Move create-bookings from BookingController to AdminController
Moved corresponding tests too.

Next objective is to update frontend
