+++ server/src/middleware/tests/BookingController.test.ts
@@ -325,136 +325,4 @@ describe("BookingController endpoint tests", () => {
       expect(res.status).toEqual(401)
     })
   })
-
-  describe("/bookings/create-bookings", () => {
-    it("should create bookings a user within the date range", async () => {
-      const bookingSlotService = new BookingSlotService()
-
-      const startDate = dateToFirestoreTimeStamp(new Date("01/01/2022"))
-      const endDate = dateToFirestoreTimeStamp(new Date("12/31/2023"))
-
-      await bookingSlotService.createBookingSlot({
-        date: dateToFirestoreTimeStamp(new Date("02/01/2023")),
-        max_bookings: 10
-      })
-
-      await bookingSlotService.createBookingSlot({
-        date: dateToFirestoreTimeStamp(new Date("03/01/2023")),
-        max_bookings: 10
-      })
-
-      // Important test case, don't return dates with no bookings
-      await bookingSlotService.createBookingSlot({
-        date: dateToFirestoreTimeStamp(new Date("01/01/2023")),
-        max_bookings: 10
-      })
-
-      const res = await request
-        .post("/bookings/create-bookings")
-        .set("Authorization", `Bearer ${adminToken}`)
-        .send({
-          startDate,
-          endDate,
-          userId: MEMBER_USER_UID
-        })
-
-      expect(res.status).toEqual(200)
-      expect(res.body.data).toHaveLength(3)
-      expect.arrayContaining([
-        expect.objectContaining({
-          users: expect.arrayContaining([
-            expect.objectContaining({ uid: MEMBER_USER_UID })
-          ])
-        })
-      ])
-    })
-
-    it("should return unauthorized error for non-admin users", async () => {
-      const startDate = dateToFirestoreTimeStamp(new Date("01/01/2023"))
-      const endDate = dateToFirestoreTimeStamp(new Date("12/31/2023"))
-
-      const res = await request
-        .post("/bookings/create-bookings")
-        .set("Authorization", `Bearer ${memberToken}`)
-        .send({
-          startDate,
-          endDate,
-          userId: undefined
-        })
-
-      expect(res.status).toEqual(401)
-    })
-
-    it("Shouldn't duplicate members in the same slot", async () => {
-      const bookingSlotService = new BookingSlotService()
-      const bookingDataService = new BookingDataService()
-
-      const startDate = dateToFirestoreTimeStamp(new Date("01/01/2022"))
-      const endDate = dateToFirestoreTimeStamp(new Date("12/31/2023"))
-
-      const slot1 = await bookingSlotService.createBookingSlot({
-        date: dateToFirestoreTimeStamp(new Date("02/01/2023")),
-        max_bookings: 10
-      })
-
-      await bookingSlotService.createBookingSlot({
-        date: dateToFirestoreTimeStamp(new Date("01/01/2024")),
-        max_bookings: 10
-      })
-
-      await bookingDataService.createBooking({
-        user_id: MEMBER_USER_UID,
-        booking_slot_id: slot1.id,
-        stripe_payment_id: ""
-      })
-
-      let res = await request
-        .post("/bookings/create-bookings")
-        .set("Authorization", `Bearer ${adminToken}`)
-        .send({
-          startDate,
-          endDate,
-          userId: MEMBER_USER_UID
-        })
-
-      expect(res.status).toEqual(200)
-      expect(res.body.data).toHaveLength(1)
-      expect.arrayContaining([
-        expect.objectContaining({
-          users: expect.arrayContaining([
-            expect.objectContaining({ uid: MEMBER_USER_UID })
-          ])
-        })
-      ])
-
-      expect(
-        (
-          await bookingSlotService.getBookingSlotsBetweenDateRange(
-            startDate,
-            endDate
-          )
-        ).length
-      ).toEqual(1)
-
-      const newEndDate = dateToFirestoreTimeStamp(new Date("01/01/2024"))
-
-      res = await request
-        .post("/bookings/create-bookings")
-        .set("Authorization", `Bearer ${adminToken}`)
-        .send({
-          startDate,
-          newEndDate,
-          userId: MEMBER_USER_UID
-        })
-
-      expect(
-        (
-          await bookingSlotService.getBookingSlotsBetweenDateRange(
-            startDate,
-            newEndDate
-          )
-        ).length
-      ).toEqual(2)
-    })
-  })
 })
